+++ ctc_ctci.c
@@ -152,7 +152,7 @@ int  CTCI_Init( DEVBLK* pDEVBLK, int argc, char *argv[] )
     if( !pWrkCTCBLK )
     {
         char buf[40];
-        MSGBUF(buf,"malloc(%lu)", sizeof(CTCBLK));
+        MSGBUF(buf,"malloc(%d)", (int)sizeof(CTCBLK));
         WRMSG(HHC00900, "E", SSID_TO_LCSS(pDEVBLK->ssid), pDEVBLK->devnum, buf, strerror(errno) );
         return -1;
     }
@@ -174,7 +174,7 @@ int  CTCI_Init( DEVBLK* pDEVBLK, int argc, char *argv[] )
     if( !pDevCTCBLK )
     {
         char buf[40];
-        MSGBUF(buf, "malloc(%lu)", sizeof(CTCBLK));
+        MSGBUF(buf, "malloc(%d)", (int)sizeof(CTCBLK));
         WRMSG(HHC00900, "E", SSID_TO_LCSS(pDEVBLK->ssid), pDEVBLK->devnum, buf, strerror(errno) );
         free( pWrkCTCBLK );
         pWrkCTCBLK = NULL;
